[PATCH, PPC 1/2] Make sure the gt- files are built for all sub-targets.
The new gt-rs6000-logue.h is common to all sub-targets in the port, so
it needs to be added for them.

It seems better to place the common target_gtfiles in the powerpc*-*-*
section, rather than duplicating them in sub-targets.  This would make it
less likely that a sub-target would be overlooked in any future file
introductions.

2019-06-26  Iain Sandoe  <iain@sandoe.co.uk>

	* config.gcc (powerpc*-*-linux*): Move target_gtfiles from here..
	(powerpc*-*-*) ... to here.



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@272708 138bc75d-0d04-0410-961f-82ee72b054a4