Allow earlyclobbers in ira_get_dup_out_num
ira_get_dup_out_num punted on operands that are matched to
earlyclobber outputs:

	    /* It is better ignore an alternative with early clobber.  */
	    else if (*str == '&')
	      goto fail;

But I'm not sure why this is the right thing to do.  At this stage
we've established that *all* alternatives of interest require the
input to match the output, so

(a) the earlyclobber can only affect other operands and
(b) not tying the registers is bound to introduce a move

The code was part of the initial commit and so isn't obviously
related to a specific testcase.  Also, I can imagine LRA makes
a much better job of this situation than reload did.  (Certainly
SVE uses matched earlyclobbers extensively and I haven't seen any
problems.)

In case this turns out to regress something important: the main
case that matters for SVE is the one in which all alternatives
are earlyclobber.

2019-07-01  Richard Sandiford  <richard.sandiford@arm.com>

gcc/
	* ira.c (ira_get_dup_out_num): Don't punt for earlyclobbers.
	Use recog_data to test for an output operand.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@272850 138bc75d-0d04-0410-961f-82ee72b054a4