[Ada] Incorrect expansion on renamings of formal parameters
This patch fixes an issue whereby a renaming of an unconstrained formal
parameter leads to spurious runtime errors; manifesting either as a
storage or constraint error due to incorrect bounds being assumed.

This issue also occurs when the renamings are implicit such as through
generic instantiations.

2019-07-03  Justin Squirek  <squirek@adacore.com>

gcc/ada/

	* sem_ch8.adb (Analyze_Object_Renaming): Add call to search for
	the appropriate actual subtype of the object renaming being
	analyzed.
	(Check_Constrained_Object): Minor cleanup.

gcc/testsuite/

	* gnat.dg/renaming13.adb, gnat.dg/renaming14.adb: New testcases.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@272982 138bc75d-0d04-0410-961f-82ee72b054a4